travel/requirements.txt
amqp==5.1.1
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==3.1.0
attrs==23.2.0
auth==0.5.3
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.7.0
blinker==1.6.3
branca==0.7.2
Brotli==1.1.0
cairocffi==1.6.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
config==0.5.1
contourpy==1.1.0
crispy==0.7.4
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.11.0
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.10
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-allauth==0.54.0
django-bootstrap4==23.2
django-braces==1.15.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-cron==0.6.0
django-embed-video==1.4.9
django-filter==23.3
django-memcache-status==2.3
django-model-utils==4.3.1
django-notifications-hq==1.8.2
django-notify==1.1.2
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
django-tinymce==3.6.1
django-twilio==0.14.3.2
django-widget-tweaks==1.4.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
env==0.1.0
environ==1.0
eventlet==0.33.3
fabio==2023.4.1
falcon==3.1.1
filelock==3.12.2
fiona==1.9.6
flake8==6.1.0
Flask==3.0.3
folium==0.16.0
fonttools==4.40.0
geopandas==0.14.4
greenlet==3.0.1
gunicorn==21.2.0
h5py==3.8.0
html5lib==1.1
idna==3.4
install==1.3.5
isort==5.13.2
itsdangerous==2.2.0
jet==1.0
Jinja2==3.1.4
joblib==1.4.2
jsonfield==3.1.0
jwcrypto==1.5.0
kiwisolver==1.4.4
kombu==5.3.1
lxml==5.2.1
mapclassify==2.6.1
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mongoengine==0.27.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.24.3
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
phonenumbers==8.13.30
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydyf==0.8.0
pyflakes==3.1.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pylint==3.1.0
pymongo==4.5.0
pyparsing==3.0.9
pypdf==4.2.0
pyphen==0.14.0
pypng==0.20220715.0
pyproj==3.6.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
python-memcached==1.59
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
shapely==2.0.4
silx==1.1.2
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
stripe==5.4.0
svglib==1.5.1
swapper==1.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.4
twilio==7.17.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.3
vboxapi==1.0
vine==5.0.0
virtualenv==20.24.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.1
whitenoise==6.5.0
wrapt==1.15.0
xhtml2pdf==0.2.15
xyzservices==2024.4.0
zopfli==0.2.3
